Add new services for our models

+++ Curtains/Curtains.Application/CurtainsService/ReviewService.cs
@@ -3,6 +3,7 @@ using Curtains.Domain.Models;
 using Curtains.Infrastructure.Interfaces;
 using AutoMapper;
 using Curtains.Application.CurtainsService.Interfaces;
+using Curtains.Infrastructure.Shared.Exceptions;
 
 namespace Curtains.Application.CurtainsServices
 {
@@ -44,6 +45,11 @@ namespace Curtains.Application.CurtainsServices
         public async Task<ReviewDTO> GetByIdAsync(int Id)
         {
             var review = await _reviewRepository.GetByIdAsync(Id);
+            if (review == null)
+            {
+                _logger.LogError("Review model is null");
+                throw new ResourceNotFoundException("Review model is null");
+            }
             var reviewDTO = _mapper.Map<ReviewDTO>(review);
             return reviewDTO;
         }
@@ -58,6 +64,11 @@ namespace Curtains.Application.CurtainsServices
         public async Task InsertAsync(ReviewDTO entity, CancellationToken cancelationToken)
         {
             var review = MapToModel(entity);
+            if (review == null)
+            {
+                _logger.LogError("Review model is null");
+                throw new ResourceNotFoundException("Review model is null");
+            }
             await _reviewRepository.InsertAsync(review, cancelationToken);
         }
 
@@ -70,6 +81,11 @@ namespace Curtains.Application.CurtainsServices
         public async Task UpdateAsync(ReviewDTO entity)
         {
             var review = MapToModel(entity);
+            if (review == null)
+            {
+                _logger.LogError("Review model is null");
+                throw new ResourceNotFoundException("Review model is null");
+            }
             await _reviewRepository.UpdateAsync(review);
         }
 
@@ -82,6 +98,11 @@ namespace Curtains.Application.CurtainsServices
         public async Task RemoveAsync(ReviewDTO entity)
         {
             var review = MapToModel(entity);
+            if (review == null)
+            {
+                _logger.LogError("Review model is null");
+                throw new ResourceNotFoundException("Review model is null");
+            }
             await _reviewRepository.RemoveAsync(review);
         }
